Use URLSearchParams for query, split to smaller files
