+++ .gitignore
@@ -59,3 +59,5 @@ typings/
 
 .vscode/
 lib/
+
+.npmrc
