+++ package.json
@@ -32,6 +32,7 @@
 	},
 	"dependencies": {
 		"@babel/runtime": "^7.10.4",
+		"fast-deep-equal": "^3.1.3",
 		"history": "^5.0.0",
 		"path-to-regexp": "^6.1.0",
 		"redux-saga": "^1.1.3"
